Move update status and download methods to main
